+++ config/dev.webpack.config.js
@@ -1,9 +1,5 @@
 const { resolve } = require('path');
 const config = require('@redhat-cloud-services/frontend-components-config');
-const TerserPlugin = require('terser-webpack-plugin');
-
-const BundleAnalyzerPlugin =
-  require('webpack-bundle-analyzer').BundleAnalyzerPlugin;
 
 const { config: webpackConfig, plugins } = config({
   rootFolder: resolve(__dirname, '../'),
@@ -37,15 +33,11 @@ plugins.push(
     }
   )
 );
-plugins.push(new BundleAnalyzerPlugin());
-
-plugins.push(new BundleAnalyzerPlugin());
 
 module.exports = {
   ...webpackConfig,
   plugins,
   optimization: {
     minimize: true,
-    minimizer: [new TerserPlugin()],
   },
 };