+++ src/Components/AffectedClustersTable/AffectedClustersTable.spec.ct.js
@@ -19,7 +19,6 @@ import {
   CHECKBOX,
   TBODY,
   TABLE,
-  CHIP,
 } from '../../../cypress/utils/components';
 import {
   DEFAULT_ROW_COUNT,
@@ -30,6 +29,9 @@ import {
   checkTableHeaders,
   checkRowCounts,
   tableIsSortedBy,
+  checkEmptyState,
+  checkNoMatchingClusters,
+  checkFiltering,
 } from '../../../cypress/utils/table';
 import {
   itemsPerPage,
@@ -51,12 +53,13 @@ const BULK_SELECT = 'clusters-selector';
 const SEARCH_ITEMS = ['ff', 'CUSTOM', 'Foobar', 'Not existing cluster'];
 const TABLE_HEADERS = _.map(AFFECTED_CLUSTERS_COLUMNS, (it) => it.title);
 
-let data = _.cloneDeep(clusterDetailData.data['enabled']);
-data.forEach(
+let values = _.cloneDeep(clusterDetailData.data['enabled']);
+values.forEach(
   (it) => (it['name'] = it['cluster_name'] ? it['cluster_name'] : it['cluster'])
 );
+const dataUnsorted = _.cloneDeep(values);
 // default sorting
-data = _.orderBy(data, ['last_checked_at'], ['desc']);
+const data = _.orderBy(values, ['last_checked_at'], ['desc']);
 
 const filtersConf = {
   name: {
@@ -111,6 +114,8 @@ describe('test data', () => {
     expect(filterData({ version: [''] })).to.have.length.gte(1);
   });
 });
+
+// TODO: when checking empty state, also check toolbar available and not disabled
 
 describe('non-empty successful affected clusters table', () => {
   beforeEach(() => {
@@ -225,6 +230,33 @@ describe('non-empty successful affected clusters table', () => {
       cy.get(MODAL).should('have.length', 1);
     });
 
+    it('checkbox can be un-clicked and all row are unselected', () => {
+      cy.ouiaId(BULK_SELECT, 'input').dblclick().should('not.be.checked');
+      // contains right text
+      cy.get('#toggle-checkbox-text').should('not.exist');
+      // checks all rows
+      cy.get(TABLE)
+        .find(TBODY)
+        .find(ROW)
+        .each((row) => {
+          cy.wrap(row)
+            .find('td')
+            .first()
+            .find('input')
+            .should('not.be.checked');
+        });
+      // bulk disabling button is not enabled
+      cy.get(TOOLBAR)
+        .find('.pf-m-spacer-sm')
+        .find(DROPDOWN)
+        .within((el) => {
+          cy.wrap(el).click();
+          cy.get('button')
+            .contains('Disable recommendation for selected clusters')
+            .should('have.class', 'pf-m-disabled');
+        });
+    });
+
     it('checkbox is unselected when a row is unselected', () => {
       cy.ouiaId(BULK_SELECT).find('input').click().should('be.checked');
       // removing one row unselects it
@@ -249,26 +281,6 @@ describe('non-empty successful affected clusters table', () => {
           cy.get('button')
             .contains('Disable recommendation for selected clusters')
             .should('not.have.class', 'pf-m-disabled');
-        });
-    });
-
-    it('checkbox unchecking removes all checks from rows', () => {
-      cy.ouiaId(BULK_SELECT).find('input').click().should('be.checked');
-
-      cy.ouiaId(BULK_SELECT).find('input').click();
-      cy.ouiaId(BULK_SELECT)
-        .find('label.pf-c-dropdown__toggle-check')
-        .contains('selected')
-        .should('not.exist');
-      cy.get(TABLE)
-        .find(TBODY)
-        .find(ROW)
-        .each((row) => {
-          cy.wrap(row)
-            .find('td')
-            .first()
-            .find('input')
-            .should('not.be.checked');
         });
     });
 
@@ -309,9 +321,7 @@ describe('non-empty successful affected clusters table', () => {
 
       cy.ouiaId(BULK_SELECT).find('input').should('be.checked');
       // contains right text
-      cy.ouiaId(BULK_SELECT)
-        .find('label.pf-c-dropdown__toggle-check')
-        .contains(`${data.length} selected`);
+      cy.get('#toggle-checkbox-text').contains(`${data.length} selected`);
       // checks all rows
       cy.get(TABLE)
         .find(TBODY)
@@ -359,6 +369,23 @@ describe('non-empty successful affected clusters table', () => {
             .should('have.class', 'pf-m-disabled');
         });
       cy.get('#toggle-checkbox-text').should('not.exist');
+    });
+
+    it('text is updated according to the number of rows selected', () => {
+      let nSelectedRows = 0;
+      // select some rows
+      cy.get(TABLE)
+        .find(TBODY)
+        .find(ROW)
+        .each((row, index) => {
+          if (index % 2 == 0 && index < DEFAULT_ROW_COUNT) {
+            cy.wrap(row).find('td').first().find('input').click();
+            nSelectedRows += 1;
+          }
+        })
+        .then(() => {
+          cy.get('#toggle-checkbox-text').contains(`${nSelectedRows} selected`);
+        });
     });
   });
 
@@ -417,12 +444,7 @@ describe('non-empty successful affected clusters table', () => {
           }
 
           // add property name to clusters
-          let sortedClusters = _.cloneDeep(
-            clusterDetailData.data['enabled'].map((it) => ({
-              ...it,
-              name: it['cluster_name'] ? it['cluster_name'] : it['cluster'],
-            }))
-          );
+          let sortedClusters = _.cloneDeep(dataUnsorted);
           // convert N/A timestamps as really old ones
           sortedClusters.forEach((it) => {
             if (it['last_checked_at'] === '') {
@@ -463,7 +485,6 @@ describe('non-empty successful affected clusters table', () => {
   });
 
   describe('filtering', () => {
-    // TODO: use filtersConf approach
     it('no chips are displayed by default', () => {
       cy.get(CHIP_GROUP).should('not.exist');
       cy.get('button').contains('Reset filters').should('not.exist');
@@ -473,37 +494,17 @@ describe('non-empty successful affected clusters table', () => {
       Object.entries(filtersConf).forEach(([k, v]) => {
         v.values.forEach((filterValues) => {
           it(`${k}: ${filterValues}`, () => {
-            const filters = {};
-            filters[k] = filterValues;
-            let sortedNames = _.map(filterData(filters), 'name');
-            filterApply(filters);
-            if (sortedNames.length === 0) {
-              // TODO check empty table view
-              // TODO headers are displayed
-            } else {
-              cy.get(`td[data-label="Name"]`)
-                .then(($els) => {
-                  return _.map(Cypress.$.makeArray($els), 'innerText');
-                })
-                .should('deep.equal', sortedNames.slice(0, DEFAULT_ROW_COUNT));
-            }
-
-            // check chips
-            for (const [k, v] of Object.entries(filters)) {
-              let groupName = filtersConf[k].selectorText;
-              const nExpectedItems =
-                filtersConf[k].type === 'checkbox' ? v.length : 1;
-              cy.get(CHIP_GROUP)
-                .contains(groupName)
-                .parents(CHIP_GROUP)
-                .then((chipGroup) => {
-                  cy.wrap(chipGroup)
-                    .find(CHIP)
-                    .its('length')
-                    .should('be.eq', Math.min(3, nExpectedItems)); // limited to show 3
-                });
-            }
-            cy.get('button').contains('Reset filters').should('exist');
+            const filters = { [k]: filterValues };
+            checkFiltering(
+              filters,
+              filtersConf,
+              _.map(filterData(filters), 'name').slice(0, DEFAULT_ROW_COUNT),
+              'Name',
+              TABLE_HEADERS,
+              'No matching clusters found',
+              false,
+              false
+            );
           });
         });
       });
@@ -512,51 +513,31 @@ describe('non-empty successful affected clusters table', () => {
     describe('combined filters', () => {
       filterCombos.forEach((filters) => {
         it(`${Object.keys(filters)}`, () => {
-          let sortedNames = _.map(filterData(filters), 'name');
-          filterApply(filters);
-          if (sortedNames.length === 0) {
-            // TODO check empty table view
-          } else {
-            cy.get(`td[data-label="Name"]`)
-              .then(($els) => {
-                return _.map(
-                  _.map(Cypress.$.makeArray($els), 'innerText'),
-                  (it) => it.replace(' \nDisabled', '')
-                );
-              })
-              .should('deep.equal', sortedNames.slice(0, DEFAULT_ROW_COUNT));
-          }
-          // check chips
-          for (const [k, v] of Object.entries(filters)) {
-            let groupName = filtersConf[k].selectorText;
-            const nExpectedItems =
-              filtersConf[k].type === 'checkbox' ? v.length : 1;
-            cy.get(CHIP_GROUP)
-              .contains(groupName)
-              .parents(CHIP_GROUP)
-              .then((chipGroup) => {
-                cy.wrap(chipGroup)
-                  .find(CHIP)
-                  .its('length')
-                  .should('be.eq', Math.min(3, nExpectedItems)); // limited to show 3
-              });
-          }
-          cy.get('button').contains('Reset filters').should('exist');
+          checkFiltering(
+            filters,
+            filtersConf,
+            _.map(filterData(filters), 'name').slice(0, DEFAULT_ROW_COUNT),
+            'Name',
+            TABLE_HEADERS,
+            'No matching clusters found',
+            false,
+            false
+          );
         });
       });
     });
 
     it('can Reset filters', () => {
-      cy.get('#name-filter').type('Not existing cluster');
+      filterApply({ name: 'Not existing cluster' });
       cy.get(TOOLBAR).find('button').contains('Reset filters').click();
       cy.get(TOOLBAR).find(CHIP_GROUP).should('not.exist');
       checkRowCounts(Math.min(DEFAULT_ROW_COUNT, filterData({}).length));
     });
 
     it('empty state is displayed when filters do not match any rule', () => {
-      cy.get('#name-filter').type('Not existing cluster');
-      // TODO check empty table view
-      // TODO headers are displayed
+      filterApply({ name: 'Not existing cluster' });
+      checkNoMatchingClusters();
+      checkTableHeaders(TABLE_HEADERS);
     });
   });
 
@@ -692,9 +673,7 @@ describe('empty successful affected clusters table', () => {
   });
 
   it('renders no clusters message', () => {
-    cy.get('#empty-state-message')
-      .find('h4')
-      .should('have.text', 'No clusters');
+    checkEmptyState('No clusters', true);
   });
 
   it('renders table headers', () => {
@@ -729,9 +708,7 @@ describe('empty failed affected clusters table', () => {
   });
 
   it('renders error message', () => {
-    cy.get('#error-state-message')
-      .find('h4')
-      .should('have.text', 'Something went wrong');
+    checkEmptyState('Something went wrong', true);
   });
 
   it('renders table header', () => {